+++ .env.default
@@ -1,0 +1,2 @@
+HELIX_RUN_QUERY_DOMAIN_KEY=aaaaa-bbbbb-ccccc-ddddd-eeeee
+DEV_PORT=3000
