+++ package.json
@@ -6,6 +6,7 @@
   "main": "src/index.js",
   "type": "module",
   "scripts": {
+    "dev": "nodemon ./src/dev.js localhost",
     "test": "c8 --100 mocha --exit -i -g 'Post-Deploy'",
     "test-postdeploy": "mocha -g 'Post-Deploy'",
     "lint": "eslint .",
@@ -58,13 +59,14 @@
     "@semantic-release/exec": "6.0.3",
     "@semantic-release/git": "10.0.1",
     "c8": "8.0.1",
-    "dotenv": "16.3.1",
+    "dotenv": "^16.3.1",
     "eslint": "8.52.0",
     "husky": "8.0.3",
     "junit-report-builder": "3.1.0",
     "lint-staged": "15.0.2",
     "mocha": "10.2.0",
     "mocha-multi-reporters": "1.5.1",
+    "nodemon": "^3.0.1",
     "semantic-release": "22.0.5"
   },
   "lint-staged": {
